requirement.txt
pudb==2019.2
einops==0.3.0
imageio==2.34.0
imageio-ffmpeg==0.4.2
omegaconf==2.1.1
test-tube>=0.7.5
torch-fidelity==0.3.0
transformers==4.38.2
pytorch-lightning==1.4.2
streamlit>=0.73.1

kornia==0.5.8
librosa==0.10.1
tqdm==4.66.2
soundfile==0.12.1
numpy==1.24.4
pandas==2.0.3
scipy==1.10.1
torchmetrics==0.5.0
torchinfo==1.8.0
-e git+https://github.com/CompVis/taming-transformers.git@master#egg=taming-transformers
